solve size_t conversion problem
git-svn-id: https://ipvs.informatik.uni-stuttgart.de/SGpp/repos/trunk@266 4eea3252-f0fb-4393-894d-40516dce545b

+++ src/sgpp/data/DataVector.cpp
@@ -4,7 +4,7 @@
 /*                                                                           */
 /* Copyright (C) 2007 Jörg Blank (blankj@in.tum.de)                          */
 /* Copyright (C) 2009 Alexander Heinecke (Alexander.Heinecke@mytum.de)       */
-/*               2007-2009 Dirk Pflueger (Dirk.Pflueger@in.tum.de)
+/*               2007-2009 Dirk Pflueger (Dirk.Pflueger@in.tum.de)			 */
 /*                                                                           */
 /* sgpp is free software; you can redistribute it and/or modify              */
 /* it under the terms of the GNU General Public License as published by      */
@@ -73,7 +73,7 @@ void DataVector::resize(size_t size)
 
 	memcpy(newdata, this->data, this->size*this->dim*sizeof(double));
 
-	for(int i = this->size; i < size; i++)
+	for(size_t i = this->size; i < size; i++)
 	{
 		newdata[i] = 0.0;
 	}
